requirements.txt
asgiref             3.7.2
decorator           5.1.1
Django              4.2.3
django-appconf      1.0.5
django-crispy-forms 2.0
django-imagekit     4.1.0
img                 2.5
package-name        0.1
pilkit              2.0
Pillow              10.0.0
pip                 23.2.1
self                2020.12.3
setuptools          68.0.0
six                 1.16.0
sqlparse            0.4.4
typing_extensions   4.7.1
wheel               0.40.0
